Replace SCrypt with BCrypt
SCrypt seems to leave dangling memory. Using the recommended memory cost
parameter of 8 will result in ~100mb of memory lingering in the JVM per
player log-in. Lowering the parameter down to 1 drops it down to 20-35mb
per player, but this still leaves a lot to be desired in comparison to
BCrypt, which doesn't leave a big block of excess memory in the JVM.