fix: combat level update doesn't update player info
The combat level update check did not include a block update for player appearance and prevented the client from showing accurate menu information involving the player (combat level of player compared to entity on right click). This change requests a player appearance block update and fixes the issue. Thanks to polar for help on this fix.