Add WorldVerificationService
This service is responsible for verifying that a player can log into the
world. The simple implementation will check if the player is online in
the current world and do other logical conditions. A more complex
implementation, such as one that supports multiple worlds, would perform
some sort of check on a separate server, such as a redis master server,
which keeps track of online players