[FIX] TileCondition functionality
Problem: `TileCondition` (used by `QueueTask.waitTile(Tile)` does not work correctly.

Symptom: `Tile` properties that are used in `Tile.sameAs(Tile)` are `val` thus cannot be changed. When a tile was passed to `TileCondition` as `src`. It seems as though the intention was for `src` to update as a reference the properties update, but the properties will never update, thus rendering `TileCondition` useless.

Solution: `TileCondition` now takes `Pawn` and `Tile`, and since `Pawn.tile` is a `var` it means that it changes as `Pawn.tile` changes, and it allows `TileCondition` to work as expected. `QueueTask.waitTile(tile: Tile)` has been changed to `QueueTask.waitTile(tile: Tile, pawn: Pawn = ctx as Pawn)`